Added a new class FormErrors
